+++ R/SklearnClassifier.R
@@ -165,11 +165,13 @@ predictPythonSklearn <- function(
   # load model
   if(attr(plpModel,'saveToJson')){
     skljson <- reticulate::import('sklearn_json')
-    modelLocation <- reticulate::r_to_py(paste0(plpModel$model,"\\model.json"))
+    ##modelLocation <- reticulate::r_to_py(paste0(plpModel$model,"\\model.json"))
+    modelLocation <- reticulate::r_to_py(file.path(plpModel$model,"model.json"))
     model <- skljson$from_json(modelLocation)
   } else{
     joblib <- reticulate::import('joblib')
-    modelLocation <- reticulate::r_to_py(paste0(plpModel$model,"\\model.pkl"))
+    ##modelLocation <- reticulate::r_to_py(paste0(plpModel$model,"\\model.pkl"))
+    modelLocation <- reticulate::r_to_py(file.path(plpModel$model,"model.pkl"))
     model <- joblib$load(os$path$join(modelLocation)) 
   }